Addressing bug.
Synchronized writing displays and metric data. Removed 'captionAll' option and just automatically caption each plot as it's created (to update displays appropriately). Updated resultsDB class so that it will update displays lines (so we can write displays multiple times -- when we first write a metric to disk and then when we create plots).


Former-commit-id: aba257f9d53fcf4410fac2d0a5b748d0508ce9a3