Updated summary stat macro.
Removed the option for a single summary stat table to have metadata across the top (one source of 'flip-flopping tables' I think, which Kem & Cathy did not like).
Updated multi-summary stat tables so that they would be more likely to be forced to have metadata on the left also (another source of flip-flopping tables).
Fixed bug which caused 'ones' in summary stat tables!


Former-commit-id: 418ec1f8818c6a282821fe14d8ad48d6af97f5cc