Put capability to plot masked points back into base spatial slicer.
Turns out this was useful for exploring opsim fields which did not receive any observations. Did not add the kwarg back into the healpix slicer plotSkyMap though, as we have a **kwargs that can just soak it up without needing to explicitly set a kwarg which we will never use there.


Former-commit-id: d44a49977410979db32ab575066c8b12cf0cb61a